Struct amethyst_dynamic_ui::layout::Layout

pub struct Layout {
    pub vars: LayoutVars,
    pub name: Option<String>,
    pub id: LayoutId,
    pub children: Vec<LayoutId>,
    pub hidden: bool,
    // some fields omitted
}

Represents a single item in the overall layout hierarchy with a bounding rectangle and an id.

Modifying any properties of this layout only stores those changes here, they won't affect the solver until this struct is passed to the solver.

Fields
vars: LayoutVarsname: Option<String>id: LayoutIdchildren: Vec<LayoutId>hidden: bool
Implementations

impl Layout

pub fn new(id: LayoutId, name: Option<String>) -> Self

Creates a new Layout.

pub fn layout(&mut self) -> &mut Self

pub fn no_container(&mut self)

Clears the container of the current Layout.

pub fn set_container<T>(&mut self, container: T) where
T: LayoutContainer + 'static,

Replaces the container of the current layout. The container is what determines what constraints will be added between this layout and it's children, as they are added, if any.

pub fn edit_left(&mut self) -> VariableEditable<'_>

pub fn edit_top(&mut self) -> VariableEditable<'_>

pub fn edit_right(&mut self) -> VariableEditable<'_>

pub fn edit_bottom(&mut self) -> VariableEditable<'_>

pub fn edit_width(&mut self) -> VariableEditable<'_>

pub fn edit_height(&mut self) -> VariableEditable<'_>

pub fn create_constraint<B: ConstraintBuilder>(
&self,
builder: B
) -> Vec<Constraint>

pub fn add<B: ConstraintBuilder>(&mut self, builder: B)

pub fn remove_constraint(&mut self, constraint: Constraint)

pub fn remove_constraints(&mut self, constraints: Vec<Constraint>)

pub fn has_constraint(&mut self, constraints: &[Constraint]) -> bool

pub fn get_constraints(&mut self) -> HashSet<Constraint>

pub fn get_removed_constraints(&mut self) -> Vec<Constraint>

pub fn get_edit_vars(&mut self) -> Vec<EditVariable>

pub fn add_child(&mut self, child: &mut Layout)

pub fn remove_child(&mut self, child: &mut Layout)

pub fn get_removed_children(&mut self) -> Vec<LayoutId>

pub fn get_children(&self) -> &Vec<LayoutId>

pub fn add_associated_vars(&mut self, vars: &LayoutVars, name: &str)

pub fn add_associated_var(&mut self, var: Variable, name: &str)

pub fn get_associated_vars(&mut self) -> Vec<(Variable, String)>

pub fn hide(&mut self)

pub fn show(&mut self)